Got rid of unnescessary classes
