--- Assets/Scripts/CreateGameInfoObj.cs
@@ -1,9 +1,0 @@
-﻿using System;
-public class CreateGameInfoObj
-    {
-        public GameInfo ConvertJSON(string JSON)
-        {
-            return GameInfo.CreateFromJSON(JSON);
-        }
-    }
-
