--- Assets/Scripts/HandleTextFile.cs
@@ -1,17 +1,0 @@
-﻿using UnityEngine;
-using UnityEditor;
-using System.IO;
-
-public class HandleTextFile
-{
-    [SerializeField]
-    string json;
-    public string WriteString(string path)
-    {
-        using (StreamReader stream = new StreamReader(path))
-        {
-            json = stream.ReadToEnd();
-        }
-        return json;
-    }
-}
